Bot-Development/Discord-Telegram-Bots/level-2/ReminderBot_MotaBhai2005/requirements.txt
discord.py==2.6.0
python-dotenv==1.0.1
pytz==2025.7
tzdata==2025.7